order_service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
httpx
sqlalchemy
PyJWT
gunicorn